refactor(mcp): move the rate-limit policy into its own file
Codacy flagged MerakiMcpClient.cs at 553 non-comment lines, over the threshold. The
signal is fair: the retry policy is a self-contained decision with no dependency on
client state, so it belongs beside the client rather than inside it. The client now
hands a tool-call delegate to MerakiMcpRateLimitPolicy.CallWithRetryAsync and keeps
only the two call sites.

Also makes the message-fragment list a static readonly field rather than allocating it
per call.

No behaviour change; all 129 affected MCP tests still green.
